2 - LLMs only/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
strands-agents[openai] 
strands-agents[litellm]>=0.1.0
strands-agents-tools>=0.1.0
strands-agents-builder>=0.1.0